sdks/python/chio-autogen/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "chio-autogen"
version = "0.1.0"
description = "AutoGen integration for the Chio protocol -- capability-scoped function registration, GroupChat role scoping, nested chat attenuation"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Backbay Industries" },
]
keywords = ["chio", "autogen", "pyautogen", "agent", "tool", "security", "capability"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]

# We depend on the classic pyautogen 0.2.x line because it exposes the
# stable ``ConversableAgent`` / ``GroupChat`` / ``register_function``
# surface referenced by the Chio roadmap. The v0.4+ ``autogen-agentchat``
# redesign pivots to an async actor model that does not expose a
# drop-in ``GroupChat`` with ``function_map``; adapting to it is tracked
# as a follow-up. pyautogen 0.2.x installs cleanly under ``uv pip`` on
# Python 3.11-3.13.
dependencies = [
    "chio-sdk-python>=0.1.0",
    "pyautogen>=0.2,<0.3",
    "pydantic>=2.5,<3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.23,<1",
    "httpx>=0.27,<1",
    "respx>=0.21,<1",
    "mypy>=1.10",
    "ruff>=0.5",
]

[tool.uv.sources]
chio-sdk-python = { path = "../chio-sdk-python" }

[tool.hatch.build.targets.wheel]
packages = ["src/chio_autogen"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = false
warn_unused_ignores = true
warn_return_any = false
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B"]
ignore = ["E501"]
